# public/openapi.yaml
openapi: 3.1.0
info:
  title: Smartkarma Gift Insight API
  version: 0.1.0
  description: |
    HTTP API for Gift Insight — Smartkarma's standalone tool for sharing
    full insights with non-clients via tokenised links.

    All endpoints are mounted under the running app's base URL
    (`APP_BASE_URL`, e.g. `https://giftinsight.smartkarma.com`).

    Authentication is cookie-based via `sk_session` (HttpOnly JWT).
    `POST /api/auth/login` and `POST /api/admin/login` set the cookie;
    `POST /api/auth/logout` clears it. Routes under the recipient flow
    (`/api/recipient/*`) and `/api/trending` are public.
servers:
  - url: '{appBaseUrl}'
    description: Application base URL
    variables:
      appBaseUrl:
        default: http://localhost:3001
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sk_session
      description: |
        HttpOnly JWT cookie set on successful sign-in. Required for any
        non-public endpoint. Some endpoints additionally require
        Insight-Provider or admin status — those return 403 if missing.
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      properties:
        error: { type: string }
        message: { type: string }
      required: [error]
...
paths:
...
  /api/admin/export:
    get:
      tags: [Admin]
      summary: Export an admin table as CSV
      description: |
        Returns one of the admin overview tables as a CSV download.
        Auth: must be signed in **and** the session email must appear in
        the `ADMIN_EMAILS` allowlist. The response carries
        `Content-Disposition: attachment` with a filename like
        `gift-insight-{table}-{YYYY-MM-DD}.csv`.

        Tables (no LIMIT — full rowset):
          - `gifters`    — every gifter aggregated
          - `authors`    — every author whose work has been gifted
          - `recipients` — every recipient who has read at least once
          - `activity`   — full activity feed (link_created / view / thanks / trial_intent)
          - `links`      — raw `gift_links` rows
          - `views`      — raw `gift_views` rows joined to their link
      parameters:
        - in: query
          name: table
          required: true
          schema:
            type: string
            enum: [gifters, authors, recipients, activity, links, views]
      responses:
        '200':
          description: CSV file.
          content:
            text/csv:
              schema: { type: string, format: binary }
        '400':
          description: Unknown `table` value.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              example: { error: invalid_table, allowed: [gifters, authors, recipients, activity, links, views] }
        '403':
          description: Not in `ADMIN_EMAILS`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              example: { error: forbidden }

  # ─────────────────────────────────────── Search ───────────────────────────────────────
...
security: []